chore: Move config to core/config
