+++ src/commands/translate/config.ts
@@ -1,6 +1,6 @@
 import {cyan, green, underline} from 'chalk';
 
-import {option, toArray} from '~/config';
+import {option, toArray} from '~/core/config';
 import {options as globalOptions} from '~/commands/config';
 
 export const NAME = 'translate';
