+++ src/core/config/index.ts
